Rename a package
