+++ ledger/participant-state-errors/src/main/scala/com/daml/ledger/participant/state/kv/errors/KVErrors.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.kvutils.errors
+package com.daml.ledger.participant.state.kv.errors
 
 import com.daml.error.definitions.ErrorGroups
 import com.daml.error.{
